Minor changes in Datafile

+++ Datafile.py
@@ -1,6 +1,4 @@
 from Answer import *
-from Answer[] import *
-from {} import *
 
 class DatafileError(Exception):
     """
@@ -46,7 +44,7 @@ class Datafile(object):
         @author
         """
         #Parameters verification:
-        if not isinstance(fileName, (str, unicode))
+        if not isinstance(fileName, (str, unicode)):
             raise DatafileError('Parameter filename must be a string or unicode.')
 
         # Setting parameters
@@ -78,33 +76,42 @@ class Datafile(object):
                 raise DatafileError('Each item on list must be an Answer object')
         self.answers = answers
 
-    def getAnswers(self):
+    def fillAnswers(self):
+        """
+         Fills the object's list containing the answers kept by the datafile with DB data.
+
+        @param Answer[] answers : List of answers that the datafile keeps.
+        @return  :
+        @author
+        """
+        cursor = MySQLConnection()
         answers = []
         answerData = cursor.execute('SELECT idAnswer FROM answer WHERE idDatafile = ' + str(self.idDatafile))
+        
         if answerData:
-            for idAnswer in answerData:
-                answers.append(Answer.pickById(idAnswer[0]))
-        return answers
+            for answerDatum in answerData:
+                answers.append(Answer.pickById(answerDatum[0]))
+        self.answers = answers
 
-    def pickById(self, idDataFile):
+    @staticmethod
+    def pickById(idDatafile):
         """
          Returns one complete Datafile object where its ID is equal to the chosen.
 
-        @param int idDataFile : Object's associated database key.
+        @param int idDatafile : Object's associated database key.
         @return Datafile :
         @author
         """
         cursor = MySQLConnection()
-        try:
-            datafileData = cursor.execute('SELECT idDatafile, fileName FROM datafile where idDatafile = ' + str(idDatafile))[0]
-        except:
+        datafileData = cursor.execute('SELECT idDatafile, fileName FROM datafile where idDatafile = ' + str(idDatafile))
+        if len(datafileData) == 0:
             return None
-        datafile = Datafile(datafileData[1])
-        datafile.idDatafile = datafileData[0]
-        datafile.setAnswers(datafile.getAnswers())
+        datafile = Datafile(datafileData[0][1])
+        datafile.idDatafile = datafileData[0][0]
         return datafile
 
-    def find(self, **kwargs):
+    @staticmethod
+    def find(**kwargs):
         """
          Searches the database to find one or more objects that fit the description
          specified by the method's parameters. It is possible to perform two kinds of
@@ -126,15 +133,11 @@ class Datafile(object):
         """
         cursor = MySQLConnection()
         datafiles = []
-        datafileData = cursor.find('SELECT idDatafile, fileName FROM datafile', kwargs)
-        try:
-            for objectData in datafileData:
-                datafile = Datafile(objectData[1])
-                datafile.idDatafile = objectData[0]
-                datafile.setAnswers(datafile.getAnswers())
-                datafiles.append(datafile)
-        except:
-            return []
+        datafileData = cursor.find('SELECT idDatafile, fileName FROM datafile ', kwargs)
+        for objectData in datafileData:
+            datafile = Datafile(objectData[1])
+            datafile.idDatafile = objectData[0]
+            datafiles.append(datafile)
         return datafiles
 
     def store(self):
@@ -145,12 +148,13 @@ class Datafile(object):
         @author
         """
         cursor = MySQLConnection()
-        if Datafile.find(fileName = self.fileName):
-            cursor.execute('UPDATE datafile SET fileName = '+ self.fileName + ' WHERE idDatafile = '+ str(self.idDatafile))
-        else: 
-            cursor.execute('INSERT INTO datafile (fileName) VALUES (' + self.fileName + ')')
-            self.idDatafile = Datafile.find(fileName = self.fileName)[0].idDatafile
-        cursor.commit()
+        if self.idDatafile is not None:
+            cursor.execute('UPDATE datafile SET fileName = "'+ self.fileName + '" WHERE idDatafile = '+ str(self.idDatafile))
+            cursor.commit()
+        else:
+            cursor.execute("INSERT INTO datafile (fileName) VALUES ('" + self.fileName + "')")
+            cursor.commit()
+            self.idDatafile = Datafile.find(fileName_equal = self.fileName)[0].idDatafile
 
     def delete(self):
         """
@@ -159,7 +163,14 @@ class Datafile(object):
         @return  :
         @author
         """
-        cursor = MySQLConnection()
-        cursor.execute('DELETE FROM datafile WHERE idDatafile = ' + str(self.idDatafile))
-        cursor.commit()
+        if self.idDatafile is not None:
+            if self == Datafile.pickById(self.idDatafile):
+                cursor = MySQLConnection()
+                cursor.execute('DELETE FROM datafile WHERE idDatafile = ' + str(self.idDatafile))
+                cursor.execute('DELETE FROM answer WHERE idDatafile = '+str(self.idDatafile))
+                cursor.commit()
+            else:
+                raise DatafileError("Can't delete non saved object.")
+        else:
+            raise DatafileError('idDatafile not defined.')
